requirements.txt
pandas>=2.0.0
numpy>=1.24.0
hmmlearn>=0.3.0
python-binance>=1.0.19
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
feedparser>=6.0.0
vaderSentiment>=3.3.2
requests>=2.31.0
praw>=7.7.0
pycryptodome>=3.20.0
psycopg2-binary>=2.9.9
